requirements.txt
astroid==1.4.9
dj-database-url==0.4.2
Django==1.10.6
django-cors-headers==2.0.2
django-vote==2.1.4
djangorestframework==3.6.2
drf-nested-routers==0.90.0
gunicorn==19.7.1
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.6.1
psycopg2==2.7.1
pylint==1.6.5
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
six==1.10.0
whitenoise==3.3.0
wrapt==1.10.10
